working on book club edit functionality.
