+++ frontend/src/index.css
@@ -112,9 +112,9 @@ hr {
 }
 
 .form__buttons {
-    display: flex;
     width: 100%;
-    gap: 10px;
+    margin: 12px 0 !important;
+    gap: 20px;
 }
 
 /* scrollbar styling */
@@ -159,6 +159,14 @@ button:active:hover {
 
 .button__auth {
     width: 100%;
+}
+
+#form__container label {
+    margin: 0 5px;
+}
+
+#form__container div {
+    margin: 8px 0;
 }
 
 #auth__container input,
